Fix for windows failure of PCHEGVX API test
The windows build of the test executable xcgsep - PCHEGVX API was seen to hang with access violation and Bad termination. This issue has been resolved. [CPUPL-7270]

+++ TESTING/EIG/pcgsepsubtst.f
@@ -670,7 +670,7 @@
          END IF
 *
 *
-         CALL IGAMX2D( DESCA( CTXT_ ), 'a', ' ', 1, 1, INT(RESULT), 1, 1, 1,
+         CALL IGAMX2D( DESCA( CTXT_ ), 'a', ' ', 1, 1, RESULT, 1, 1, 1,
      $                 -1, -1, 0 )
          IF( RESULT.NE.0 )
      $      GO TO 160
@@ -803,7 +803,7 @@
 *
 *     All processes should report the same result
 *
-      CALL IGAMX2D( DESCA( CTXT_ ), 'a', ' ', 1, 1, INT(RESULT), 1, 1, 1, -1,
+      CALL IGAMX2D( DESCA( CTXT_ ), 'a', ' ', 1, 1, RESULT, 1, 1, 1, -1,
      $              -1, 0 )
 *
   160 CONTINUE
